Fix #2066: Emit `x.isInstanceOf[C]` as `x instanceof C` when possible.
When the rhs of an `isInstanceOf` refers to a *class* `C` (as
opposed to a trait), it is semantically correct to compile it as
`x instanceof $c_C`. It turns out that this is much faster.

In fastOpt mode, the optimization also affects `asInstanceOf`s.

In fullOpt mode, most instance tests come from pattern matching,
so effectively this optimization is very effective on pattern-
matching-intensive workloads.

Benchmarks of the bootstrapped Scala.js optimizer while optimizing
the test suite show speedups of up to 47% in fastOpt and 40% in
fullOpt, with the maximum for the optimizer core.