Fix #1201: Correct regex support.
Previously, `java.util.regex.Pattern` was implemented without much
concern for correctness wrt. the Java semantics of regular
expressions. Patterns were passed through to the native `RegExp`
with minimal preprocessing.

This could cause several kinds of incompatibilities:

- Throwing `ParseError`s for features not supported by JS regexes,
- Or worse, silently compile with different semantics.

In this commit, we correctly implement virtually all the features
of Java regular expressions by compiling Java patterns down to JS
patterns with the same semantics.

This change introduces a significant code size regression for code
bases that were already using `Pattern` and/or Scala's `Regex`.
Therefore, we went to great lengths to optimize the compiler for
code size, in particular in the default ES 2015 configuration. This
means that some code is not as idiomatic as it could be. The impact
of this commit on a typical output is approximately 65 KB for
fastOpt and 12 KB for fullOpt.

The `README.md` file contains an extensive explanation of the
design of the compiler, and of how it compiles each kind of Java
pattern.

In addition to fixing the mega-issue #1201, this commit fixes the
smaller issues #105, #1677, #1847, #2082 and #3959, which had been
closed as duplicates of #1201.