Regex: Use the native 'd' flag to compute indices when available.
That feature is enabled by a flag, and not by default, because of
performance concerns. JS engine implementers have shown that
enabling the feature has a significant performance impact.
Therefore, we only enable it the first time group indices are
required for any given `Pattern` instance. This means that the
first match for which group indices are required will be run twice,
but this is probably the better trade-off.