Directly call ES2015 builtins instead of storing them in a var.
Previously, when emitting ES2015, we stored some builtins that we often
use in global var defs:

  var $imul = Math.imul;
  var $fround = Math.fround;
  var $clz32 = Math.clz32;

It turns out that engines, especially V8, do a better job at optimizing
direct calls such as `Math.fround(x)` rather than indirect calls such
as `$fround(x)`. And that despite the fact that `$fround` can be
"proven" to be constant, contrary to `Math` or `Math.fround`.

Now, when emitting ES2015, we directly emit calls to `Math.fround(x)`
instead of storing `$fround`. The same applies to `Math.imul` and
`Math.clz32`.

We were already applying that strategy to `Object.is` (also ES2015+)
and `Object.getOwnPropertyDescriptors` (ES2017+). We consolidate the
logic used for all five of these polyfillable builtins.