Refactoring: Focus internal module ID generation in InternalModuleIDGenerator.
It is an evolution of `ModuleIDs` into a class that stores more
information about what it needs to avoid, so that we don't have to
repeat it at use site.

It also concentrates all the logic about internal module ID
generation, which will allow the future fix for case-insensitive
filesystems to be more localized.