Fix #4855: Avoid case-insensitive clashes for internal module IDs.
When an internal module contains a single class, we name it after
the class name. This is supposed to give unambiguous module IDs.
However, it is not good enough on case-insensitive file systems,
as we can have two classes whose names differ only in case.

To avoid the issue, we prefix uppercase ASCII characters with
a '-' and all non-ASCII characters with a '-' and their code
point value.

We have a similar issue for digest-based module IDs. Their
'internal-' prefix must also be protected for collision in a
case-insensitive way. For that, we can directly compare with
prefixes of the IDs `equalsIgnoreCase`.