Avoid propagating global refs that should not be tracked to the top.
Normally, unless we track all global refs for the benefit of GCC,
we prune away non-dangerous global refs as soon as possible, to
preserve empty sets. This is done in two ways:

* Within `FunctionEmitter`, we track everything for local purposes,
  but then prune the non-dangerous global refs when escaping back
  to `ClassEmitter`.
* Everywhere else, we don't track non-dangerous global refs in
  the first place.

This policy was broken in two places, because of helper code that
is called both from inside and outside `FunctionEmitter`. These
functions now receive a flag to indicate whether they should keep
only the global refs that need to be tracked (for outside
`FunctionEmitter`) or all of them (for inside it).