Under Minify, inline `VarDef`s used only once when we can.
During the optimizer, when emitting a `VarDef(x, ..., rhs)`, we
try to inline it if it has been used exactly once. In order to do
that, we look at the `body` in which it will be available, and
replace its only occurrence if it occurs in the first evaluation
context following only pure subexpressions.

See the long comment in the code for more details.