Make the optimizer type-preserving by inserting Casts when necessary.
Previously, the optimizer generated IR that was ill-typed, in the
sense that it would not pass the IR checker (even if it were
generalized to understand intrinsics and record types).

This happened when the receiver of a method call had a more generic
type than an inlined method. That can be the case when only there
is effectively only one possible target method for a virtual call,
or when there are multiple targets but we can apply multi-inlining.

Producing ill-typed IR is not a big deal for JavaScript, although
we had to care about some details like proper unboxing of chars.
In order to apply the optimizer to the Wasm backend, however, we
need well-typed IR to be able to produce well-typed Wasm.

In this commit, we fix the issue by introducing a new transient,
`Cast`. It performs a true, unchecked cast, doing nothing but
reassigning another type to an expression. It behaves like an
unchecked `AsInstanceOf`, except it does not convert `null` to the
zero of primitive types.

When inlining, we now introduce casts (along with explicit null
checks) to adapt the receiver to the target method. For single
inlining, it is straightforward. For multi-inlining, choosing a
correct target type, and generally producing well-typed IR,
required to significantly change the logic. Previously, we picked
an arbitrary target among the possible candidates and inlined it
in the normal way. We cannot do this anymore, since the `This`
type of any one possible target may not represent what is valid
for all the targets. Instead, we now deeply inspect the shapes
we know how to multi-inlining, and recreate calls on the fly for
the method that they call.

The explicit introduction of casts at the inlining level, along
with explicit null checks, made the logic to refine receiver
types in `withBindings` redundant. In order to completely get
rid of it without any regression, we had to make `checkNotNull`
more powerful to preserve the "it is now not-null" information
even when `nullPointers` are unchecked.

Interestingly, making the above changes retains more information
even within the optimizer. That results in fewer useless
temporary variables in the generated code. The changes are
therefore beneficial even for the JavaScript backend.

Finally, since we have proper casts, we can simplify
`AsInstanceOf`s that never require unboxing into `Cast`s, when
`asInstanceOfs` are unchecked. This changes almost nothing to the
generated code, except replacing some `$uC(expr)` by `expr.c`.