Wasm: Use custom JS helpers for almost all JS interop.
Previously, we used a small set of fixed JS helpers to implement
all the Wasm-to-JS interop nodes. For example, there was a unique
`jsMethodApply` helper, used for all `JSMethodApply` nodes. We also
used a single entry point to read all `JSGlobalRef`s, using a JS
`new Function("...")` to read global variables.

This strategy, while having the advantage of doing all the
codebase-dependent work on the Wasm side, was extremely slow.
Imagine that calling `Math.cos(x)` had to:

1. load the constant string `"Math"`,
2. use `jsGlobalRefGet("Math")`, using a `new Function` to read that,
3. load the constant string `"cos"`,
4. use `jsNewArray` and `jsArrayPush` to create a JS array with `x`
   as argument (plus a `bD` call to box `x`!),
5. use `jsMethodApply`, which uses a `...spread` operator to
   re-expand the one-element array of arguments.

A unique call like that in a more-or-less tight loop would destroy
any hope of good performance.

---

We now use a completely different strategy. For every Wasm-to-JS
interop node, `FunctionEmitter` creates a dedicated JS helper. It
is registered in the context and will be inserted in the generated
`main.js` file.

We build on a unified toolkit to generate efficient JS helpers and
managing their inputs: `CustomJSHelperBuilder`. It manages the
evaluatation of Wasm values on one side, then gives them to a
dedicated JS function body. Boxing is fused into the JS function
call through the built-in Wasm-to-JS interop features when
possible. Moreover, literals and JS global refs are
"constant-folded" into the JS function altogether.

For the above example scenario, we end up with a helper of the form
`(x) => Math.cos(x)`, typed in Wasm as a `[f64] -> [anyref]`. We
can call it directly with a `double` value, without explicit call
to `bD`, without JS array, and without string values representing
`"Math"` nor `"cos"`. This is almost as good as it gets. The only
non-optimal aspect is that we still return an `anyref`, which we
have to unbox later with a call to `uD`. Improving that last point
is left for future work.

---

We use the same mechanism to build dedicated helpers for `Closure`s,
top-level exported defs and for non-native JS classes.

---

There are two things that we can still improve in the future
without too much effort:

* Handling of the arguments to the super constructor calls in
  non-native JS classes, for which we could leverage Wasm functions
  with multiple return values.
* Move the management of the `Symbol`s for private JS fields to the
  JS side of things, instead of passing them from Wasm every time.