Wasm: Translate string types to `(ref null? extern)`.
Instead of `(ref null? any)`.

This causes a lot of additional conversions between `extern`s and
`any`s when boxing and unboxing, without any added value. This is
only a stepping stone for the next commit. In the latter, we will
start using JS string builtins, which manipulate `extern`s. We will
want strings to already be typed as `extern`s for all the string
manipulations.