Refactor: Manipulate "conditional features" instead of "instantiated classes".
Previously, the `Infos` and more importantly the Emitter's
`GlobalKnowledge` manipulated special "instantiated classes", which
conditioned whether some functions where emitted or not. This
naming matched what we were actually looking at from the `Analysis`,
but not what we were doing it *for*.

Now, we instead talk about "condition features", which directly
target the functions that are conditionally emitted. They are still
inferred from the `Analysis`, but with an added constraint on the
specific constructor used to instantiate the classes.

We also refactor the `Infos` builder to highlight this strategy as
well, even though it ends up doing the same thing in terms of
analysis.