Fix #5044: Handle the IllegalArgumentExceptions in jlr.Array.newInstance.
The case of an empty array of dimensions for the multi-dimensions
overload is handled in user-space. The case of `classOf[Unit]` is
handled in the backends, since not doing so leaves the door open
for user-space IR to break the fundamental invariant that there is
no array of `void`.

In the JS backend, we follow the strategy of int/long div/mod to
emit the target method only if `IllegalArgumentException.<init>()`
is reachable. In the Wasm backend, since we do not have that
infrastructure, for now we add an unconditional dependency on that
constructor, like we do for `jl.Class.<init>()`.