Wasm: Lower block with parameters to blocks without params.
Currently, wasm-opt does not support block types with parameters.
In order to pipe our output to wasm-opt, we therefore add a pass
that lowers `block`s with parameters.

For simplicity, we only implementation the transformation for
`block`s (and not other block-type instructions). We also assume
that when there are parameters to lower, there are not multiple
result types at the same time. These assumptions allows the
transformation to remain pretty straightforward; currently, our
codegen only generates code that fulfills those assumptions.