Refactoring: Generic notion of `CheckingPhase` for the IR checkers.
It was bad enough that we had two boolean options for deciding what
IR is allowed or not between phases. There was already one of the
four combinations that was invalid.

With the introduction of the desugarer, we had *three* such boolean
options. Only 4 out of their 8 combinations were valid.

We now introduce a generic concept of `CheckingPhase`, which knows
what conditional IR features it *accepts*. This makes a lot of
things clearer.