Refactoring: Generic notion of CheckingPhase for the IR checkers.
We previously had two boolean options for deciding what IR is
allowed or not between phases. Of the four possible combinations
of values, one was actually invalid.

With the introduction of the desugarer, there would have been
*three* such boolean options. Only 4 out of their 8 combinations
would have been valid.

We now introduce a generic concept of `CheckingPhase`. Some IR
features are allowed as input of each `CheckingPhase`. Since the
`ClassDefChecker` and `IRChecker` must agree on what features are
valid when, we factor that logic into `FeatureSet`. It represents
a set of logical features, and can compute which set is valid for
each phase.