Wasm: Implement fmod on the Scala side.
This avoids the JS call overhead for that "primitive" operation.

As can be seen in the implementation, `fmod` is far from a
primitive, though. It requires an involved algorithm in software.

We took an MIT implemention of `fmod` written in Rust, generic
in the bit-width, and translated it to Scala. The "generic"
aspect is turned into a big copy-pasted blob between the `Float`
and `Double` versions, with some "configuration" at the top.

We put that implementation in the linker private library, in a
new object `WasmRuntime`. The Wasm backend compiles `Float_%` and
`Double_%` as calls to those runtime functions.