Emit a different loader file for targetPureWasm.
This loader file does not expose any of the JS interop features.
This way, we can use the standard `run` (and later `test`) sbt
commands, while verifying that are indeed not touching any JS.

Moreover, the build sets up the most limited flags for Node.js
based on the linker config, so we also don't have to take care of
that while testing.