Implement conversion of Long to String.
Previously, `LongTest.testToString()` appeared to work because all
computations were constant-folded. We made the test more robust
wrt. constant-folded, to make sure to surface the issues even when
the optimizer is enabled.

We implement the long-to-string conversion as `ltoa` in the Wasm
backend. We make `itoa` delegate to `ltoa`, so that we do not
duplicate code.