Implement TestAdapter for pure Wasm using some essential JS imports
This commit implements a TestAdapter for a pure Wasm environment using two additional essential JavaScript imports: send and init.
Each of these imports is a proxy for the send and init methods of the scalajsCom object, which is defined in the JavaScript global.

https://github.com/scala-js/scala-js-js-envs/blob/d0683564ef64c86731e1037a30d5bec13c95b5fd/nodejs-env/src/main/scala/org/scalajs/jsenv/nodejs/ComSupport.scala#L289-L307

Also, add an example project for running tests using this TestAdapter implementation.
Once we start using this TestAdapter in the test suite, we can remove the project.