Implement regular expressions.
The strategy is to reimplement a regex engine that matches the
semantics of `js.RegExp`. We reuse the `PatternCompiler` to deal
with all the complicated Java patterns. The engine only needs to
handle the simplified regexes emitted by our `PatternCompiler`.

We abstract away the underlying engine as an abstract `Engine`.
On JS, it is zero-cost forwarder to `js.RegExp`. On Wasm without
a JS host, it is an entirely separate implementation.

The implementation closely follows the ECMAScript spec text. The
core of the algorithm is a CPS-based exploration of alternatives
with backtracking. This is not the most efficient possible, but
for now, it has the benefit of remaining pretty simple. It is also
easier to convince oneself that it is correct.

One critical exception to that strategy is the canonicalization of
`CharSet` through the simple case folding function. In order to
get good performance, we precompute (during `WasmEngine.compile`)
the mapping of a `CharSet` through the simple case folding
function. This is a fairly intrincate algorithm in
`CharSet.mapBySimpleCaseFolding()`.