identityHashCode on boxed float and double should use Double.hashCode
identityHashCode on a FloatBox, DoubleBox, and IntegerBox, should perform the hashCode function on respective class, such as Double.hashCode.
However, previously, in pure Wasm, the implementation of the boxed class was to use the value taken from the lastIDHashCode, so that hashCode functions of two java.lang.Double classes with the same value returned different values.