Eliminate FloatBox class
f32 can be boxed into a DoubleBox. This reduces the number of ref.test against Box class in runtime type check.
Similarly, i32 (especially, a value that does not fit in i31ref) can be boxed into the DoubleBox to eliminate the IntegerBox, but so we leave it as is for the time being, since it's not obvious to improve performance.