Remove all the fast paths from RuntimeLong division.
Now that the worst case isn't so bad anymore, we remove the fast
paths for small values of the numerator.

It does make them slower. However, it also makes division more
predictable: the time is now dependent only on the divisor, which
should be more stable in most applications. It also reduces the
overall code size.