Wasm friendly parse float implementation
This commit implements a Bellerophon algorithm from
"How to Read Floating Point Numbers Accurately" by William D. Clinger.

The Bellerophon algorithm efficiently finds the best binary
approximation for a decimal number `f * 10^e`.
While an iterative approach like AlgorithmM (defined in the same paper)
can find the best approximation by repeatedly using slow
BigInteger arithmetic, Bellerophon provides non-iterative approach.

The basic idea is, instead of directly computing the best approximation
for `f * 10^e` it uses higher-precision (p-bit (64-bit) significand)
floating-point numbers to approximate `f` and `10^e` separately.
Then multiplies these approximations and analyzes the
error when multiplying and rounding the result to the target
n-bit (n-bit significand) precision.

In some cases, the calculated approximation isn't the best, and in that case
the algorithm falls back to AlgorithmR.
This algorithm takes a good appoximation as a starting point,
and iteratively finds the best approximation.
Since the calculated approximation by Bellerophon is very close to the best
it converges in just a few steps.

For extreme values that would overflow to infinity or underflow to a
subnormal number or zero, neither the Bellerophon nor AlgorithmR does work.
In these cases, the algorithm falls back to the original, slower AlgorithmM
with the modification mentioned in section 8 of the paper.
> With IEEE arithmetic, for example, a denormalized result
> may be required. Denormalized results can be generate
> by a modified form of AlgorithmM that terminates immediately
> when the minimum exponent is reached.

In future, we might want to migrate to Eisel-Lemire algorithm?