Wasm: Store the type data of primitive arrays directly in globals.
Most arrays are arrays of primitive or of `jl.Object`. It is
wasteful to call `arrayTypeData` to load their vtable every time,
especially since it is required for every instantiation.

Instead, we now eagerly create the type data for these base types
and store them in immutable globals (like the vtabls of regular
classes).

One apparent downside is that the `arrayOf` field of the primitive
type data won't be bound to those. It cannot be done in the
initialization because that would be a cycle in global init order.
We *could* initialize them in the `start` function, though.

It turns out a better idea is to *forbid* calling `arrayTypeData`
when the base type is a primitive or `jl.Object`. That means their
`arrayOf` field will never be used, and always stay `null`. This
is acceptable because it's easy to meet that requirement. We thus
rename `arrayTypeData` to `specificArrayTypeData` to make that
clearer.

Since `specificArrayTypeData` can never see a primitive base
anymore, its code is simplified. It always uses the same `clone`
method. Likewise, we can simplify `typeDataName`, because we can
eagerly store the names of arrays of primitives, which do not
follow a standard pattern.

Overall, though there is a subtlety about the new spec for
`specificArrayTypeData`, I feel this is overall a simplification,
in addition to producing better code.