Wasm: Implement `@JSExport` support when using custom descriptors.
When using custom descriptors, we now bind the vtables to JS
prototypes that carry the `@JSExport`ed members. We only create
separate prototypes for classes that introduce at least one
exported member. Otherwise, we reuse the same prototype as the
parent class.

In this commit, we use custom JS helpers to configure all the JS
prototypes. In the future, we should do most of that using the
`configureAll` builtin introduced by the custom descriptors
proposal. Only methods with a `...rest` param should still need
custom JS code. That is an optimization left for later.