Wasm: Implement `@JSExport` support when using custom descriptors.
When using custom descriptors, we now bind the vtables to JS
prototypes that carry the `@JSExport`ed members. We only create
separate prototypes for classes that introduce at least one
exported member. Otherwise, we reuse the same prototype as the
parent class.

In this commit, we use custom JS helpers to configure all the JS
prototypes. In the future, we should do most of that using the
`configureAll` builtin introduced by the custom descriptors
proposal. Only methods with a `...rest` param should still need
custom JS code. That is an optimization left for later.

As of this writing, the latest version of Node.js (25.1.0) uses an
earlier version of the proposal, where the actual prototype objects
had to be wrapped in a `DescriptorOptions` object. We currently use
code paths to handle that earlier spec for testing purposes. We
also (and mainly) support the newer version, which we can manually
test with Chrome for now.