Port RyuDouble
Implement the Ryu for converting binary floating-point numbers the decimal string representation, as described in "Ryu: fast float-to-string conversion" by Ulf Adams (PLDI 2018).
https://dl.acm.org/doi/10.1145/3192366.3192369

This port is based on the reference Java implementation from https://github.com/ulfjack/ryu with some comments and changes.

We use `RyuDouble` both for Float and Double because of Scala.js spec.

We do not constant fold on JVM on purpose because
the rounding behavior of Double.toString differs between versions before and after JDK 19.
For instance, in JDK 21, `Double.toString(double)` specifies "round to even" for boundary cases,
whereas this specific boundary behavior seems not specified in JDK 17.
https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/lang/Double.html#toString(double)
https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Double.html#toString(double)