Move the logic for "exactness" of primitives into `typeTestResult`.
`typeTestResult` was now the only place the `isExact` flag of a
`RefinedType` was even looked at for primitive base types.

Instead of trying to assign `isExact` for some primitive types for
every created `RefinedType`, we move that logic to `typeTestResult`.
In there, we can even do a more thorough analysis, because we also
know the target `testType`.

This results in slightly better code for some type tests and casts
involving primitive numeric types.