Never mark bridge/mixin methods as noinline.
They copy the annotations of their target methods, but there is
never a good reason to prevent inlining the bridges themselves.

Also, remove an obsolete list of exclusions. `ControlContext` was
part of the continuations library, which has been removed from
Scala years ago. And the methods of `List` have not had the
`@noinline` anntation since Scala 2.12.0 (previously it was there
to work around a JVM inliner bug, apparently).