Introduce PreTransIf in the optimizer.
This fixes an inconsistency in how we deal with folding. For all
other trees that we `fold`, we are doing so on the `PreTransform`s.
This is good, because folding sometimes entirely dce's
subcomputations. If they are `PreTransform`s, that can lead to vals
not being `used` at all, which allows them to disappear in turn.

For `If`s, we were previously doing folding on the fully transformed
`Tree`s. This resulted in suboptimal folding, notably when `hi` and
`lo` parts of a split long were involved. Indeed, those are
`PreTransLocalDef`s prior to being "finished", but after they become
slightly more complex than a `VarRef`, which defeated some folding
logic.