WiP New encoding for the Unicode database.
The Unicode database used by `jl.Character.getType` and the various
`jl.Character.isXYZ` methods is now entirely generated by an sbt
task.

The relevant methods are isolated in a dedicated object
`UnicodeData`, to allow DCE without slowing down access to them.

More importantly, we use a new encoding that is both more code
size-efficient and run-time-efficient.

For run-time, we store each range as a pair `(firstCP, data)` in a
single array. The data contain the type and a bit-field of
properties that cannot directly be derived from the type. A single,
efficient method `dataHasAnyFlag` can test for a set of types and
a set of properties. For example, `isAlphabetic` returns `true`
when either the type is any of the five major `XYZ_LETTER` types
or `LETTER_NUMBER`, or when the property `OtherAlphabetic` is on.

For code size, we compress each pair as a single `Int`. It contains
the `data` (minus "computed" fields of the bit-field that are
computed at load time) as well as a `diff` from the previous range's
`firstCP` to the current one. Some diffs are too large to fit in the
available bits, and are special-cased.

Since the logic is bit hackery-heavy, we also generate exhaustive
test data with straightforward logic.