New encoding for the Unicode database.
We now generate the Unicode database used by `jl.Character.getType`
and the various `jl.Character.isXYZ` methods in an sbt task.

The relevant methods are isolated in a dedicated object
`UnicodeData`, to allow DCE without slowing down access to them.

More importantly, we use a new encoding that is both more code
size-efficient and run-time-efficient.

For run-time, we store each range as a clever bit field:

  firstCP (21) ++ props (5) ++ alternatingTypes flag (1) ++ type (5)

which, remarkably, adds up to 32 bits.

There are 5 boolean properties that cannot be derived from the type
field, but that are used in conjunction with the type field to
implement some of the `isXYZ` methods. A single, efficient method
`dataHasAnyFlag` can test for a set of types and a set of properties.
For example, `isAlphabetic` returns `true` when either the type is
any of the five major `XYZ_LETTER` types or `LETTER_NUMBER`, or when
the property `OtherAlphabetic` is on.

For code size, we further compress the `firstCP` field as a *diff*
from the previous code point. We expand them in-place during the
constructor of `UnicodeData`.

Two `isXYZ` methods, namely `isIdeographic` and `isMirrored`, only
rely on a boolean property, and not on the type. Since those
properties are not even very correlated to the ranges used for the
main data, they receive their independent array of ranges.

Since the logic is bit hackery-heavy, we also generate exhaustive
test data with straightforward logic.